ConnectionManager interface
